+++ yazi-dds/src/body/mod.rs
@@ -7,7 +7,7 @@ mod custom;
 mod hey;
 mod hi;
 mod hover;
-mod move_event;
+mod move_;
 mod rename;
 mod yank;
 
@@ -18,6 +18,6 @@ pub use custom::*;
 pub use hey::*;
 pub use hi::*;
 pub use hover::*;
-pub use move_event::*;
+pub use move_::*;
 pub use rename::*;
 pub use yank::*;
